requirements.txt
# ── Web framework ─────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Config & validation ────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── Async file I/O ─────────────────────────────────────────────────
aiofiles==23.2.1

# ── LangChain + LangGraph (agent orchestration) ────────────────────
langchain==0.2.5
langchain-community==0.2.5
langchain-ollama==0.1.3
langgraph==0.1.19

# ── Vector store + embeddings ──────────────────────────────────────
faiss-cpu==1.8.0
sentence-transformers==2.7.0
langchain-huggingface==0.0.3

# ── Financial document parsing ─────────────────────────────────────
pdfplumber==0.10.3
pypdf==4.2.0
python-docx==1.1.2
docx2txt==0.8

# ── Knowledge graph ────────────────────────────────────────────────
neo4j==5.19.0

# ── Authentication ─────────────────────────────────────────────────
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# ── Evaluation ────────────────────────────────────────────────────
ragas==0.1.10
datasets==2.19.0

# ── Testing ────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

celery==5.3.6
redis==5.0.4
websockets==12.0